+++ compiler/include/view.h
@@ -4,13 +4,13 @@
 #include "baseAST.h"
 #include "chplenum.h"
 
-extern BaseAST* ast(long id);
+extern BaseAST* ast(int id);
 extern void print_view(BaseAST* ast);
 extern void print_view_noline(BaseAST* ast);
-extern void iprint_view(long id);
+extern void iprint_view(int id);
 extern void nprint_view(BaseAST* ast);
 extern void nprint_view_noline(BaseAST* ast);
-extern void mark_view(BaseAST* ast, long id);
+extern void mark_view(BaseAST* ast, int id);
 
 extern void print_symtab(BaseAST* ast);
 extern void nprint_symtab(BaseAST* ast);
